refactor(age,auth): extract age + auth to src/modules — KAN-415 D6
Moves the two modules the plan deliberately sequenced together, because
their combined footprint IS the signup gate's footprint:

  src/lib/age/{age-service,didit,provider-gate,record-declaration,
               self-declaration}.ts  -> src/modules/age/
  src/lib/auth/post-login-redirect.ts -> src/modules/auth/

18 files had their import specifiers rewritten. No behaviour change: every
moved file is byte-identical, and src/modules/auth/post-login-redirect.ts
imports the age module exactly as it did before (that coupling is why the
two move in one commit rather than two).

THE SIGNUP GATE IS THE POINT OF THIS STEP

.github/signup-surface.paths listed src/lib/age/** and src/lib/auth/**.
KAN-419 demonstrated that moving them without rewriting the manifest makes
the gate report RESULT: CLEAN — exit 0, no warning, no log — while account
creation goes unproven into staging, because the daily soak deliberately
never exercises signup. The manifest is rewritten in THIS commit, so no
window exists in which the gate is disarmed.

Deviation from KAN-419 §4, deliberately: that section recommends landing a
UNION manifest (old and new globs together) first, then shrinking. The
union exists to cover a move that spans multiple commits. This move is
atomic, so there is no interval to cover, and a union would leave
src/lib/age/** matching no tracked file — which CTL-035 would correctly
fail. Atomic is strictly safer here; the union remains right for any
extraction that cannot be done in one commit.

Every path-keyed artefact updated in the same commit (the KAN-428
extraction DoD):
  - .github/signup-surface.paths   both globs, with the reason recorded
  - modules.json                   15 paths (age + auth already existed
                                   as planned modules, pointing at src/lib)
  - env-access-baseline.json       CTL-037 keys files by path; didit.ts
                                   reads env, so a stale key would fail the
                                   ratchet in BOTH directions
  - tests/support/source-path-seeds.ts + regenerated SRC manifest

THE MANIFEST TRAP FIRED AGAIN, AND TSC CAUGHT IT

Moving these dropped SRC.providerGate and SRC.postLoginRedirect for the
reason documented in source-path-seeds.ts: the generator keeps only
literals that still resolve, and no test carried the new paths yet. This
time `tsc --noEmit` failed the build — the typed manifest turns the whole
class into a COMPILE error for .ts tests, which is the earliest place it
can surface. Both keys are now seeded.

Test estate unchanged at 3544 (273 suites); tsc clean.

Refs: KAN-415, KAN-413, KAN-419, KAN-428
