+++ env-access-baseline.json
@@ -66,7 +66,7 @@
     "src/lib/affiliate/link-service.ts": [
       "SOVRN_API_KEY"
     ],
-    "src/lib/age/didit.ts": [
+    "src/modules/age/didit.ts": [
       "DIDIT_API_BASE",
       "DIDIT_API_KEY",
       "DIDIT_WEBHOOK_SECRET",
