+++ src/app/auth/confirm/route.ts
@@ -1,7 +1,7 @@
 import { createClient } from '@/modules/platform/supabase-server';
 import { NextResponse } from 'next/server';
 import type { EmailOtpType } from '@supabase/supabase-js';
-import { resolvePostLoginRedirect } from '@/lib/auth/post-login-redirect';
+import { resolvePostLoginRedirect } from '@/modules/auth/post-login-redirect';
 
 /**
  * BUGS-50 — email confirmation via the token-hash (verifyOtp) flow.
