+++ src/app/confirm-age/actions.ts
@@ -10,8 +10,8 @@ import { redirect } from 'next/navigation';
 import { headers } from 'next/headers';
 import { createClient } from '@/modules/platform/supabase-server';
 import { env } from '@/modules/platform/env';
-import { recordAgeDeclaration } from '@/lib/age/record-declaration';
-import { resolvePostLoginRedirect } from '@/lib/auth/post-login-redirect';
+import { recordAgeDeclaration } from '@/modules/age/record-declaration';
+import { resolvePostLoginRedirect } from '@/modules/auth/post-login-redirect';
 
 export async function confirmAge(formData: FormData) {
   const supabase = await createClient();
