+++ src/app/confirm-age/page.tsx
@@ -15,7 +15,7 @@ import Link from 'next/link';
 import Image from 'next/image';
 import { redirect } from 'next/navigation';
 import { createClient } from '@/modules/platform/supabase-server';
-import { hasDeclaredAge } from '@/lib/age/record-declaration';
+import { hasDeclaredAge } from '@/modules/age/record-declaration';
 import { confirmAge, declineAge } from './actions';
 
 export const metadata = {
