+++ src/app/verify-age/actions.ts
@@ -9,8 +9,8 @@
 import { redirect } from 'next/navigation';
 import { createClient } from '@/modules/platform/supabase-server';
 import { env } from '@/modules/platform/env';
-import { createAgeSession } from '@/lib/age/didit';
-import { profileIdForUser, setProfileAgeStatus } from '@/lib/age/age-service';
+import { createAgeSession } from '@/modules/age/didit';
+import { profileIdForUser, setProfileAgeStatus } from '@/modules/age/age-service';
 
 export async function startAgeVerification(): Promise<void> {
   const supabase = await createClient();
