+++ tests/support/source-paths.json
@@ -147,9 +147,6 @@
     "statusPage": "src/app/status/page.tsx",
     "components": "src/components",
     "lib": "src/lib",
-    "providerGate": "src/lib/age/provider-gate.ts",
-    "selfDeclaration": "src/lib/age/self-declaration.ts",
-    "postLoginRedirect": "src/lib/auth/post-login-redirect.ts",
     "authBearer": "src/lib/convene/auth-bearer.ts",
     "microsoft": "src/lib/convene/calendar/microsoft.ts",
     "env": "src/lib/convene/env.ts",
@@ -169,6 +166,9 @@
     "modules": "src/modules",
     "gate": "src/modules/access/gate.ts",
     "accessPipeline": "src/modules/access/pipeline.ts",
+    "providerGate": "src/modules/age/provider-gate.ts",
+    "selfDeclaration": "src/modules/age/self-declaration.ts",
+    "postLoginRedirect": "src/modules/auth/post-login-redirect.ts",
     "entitlementsService": "src/modules/features/entitlements-service.ts",
     "fileMagicBytes": "src/modules/guards/file-magic-bytes.ts",
     "sanitise": "src/modules/guards/sanitise.ts",
